Addon-docs: Port Vue to ArgsTable
