+++ addons/docs/src/lib/sbtypes/utils.ts
@@ -1,0 +1,2 @@
+const QUOTE_REGEX = /^['"]|['"]$/g;
+export const trimQuotes = (str: string) => str.replace(QUOTE_REGEX, '');
